requirements.txt
cryptography
argon2-cffi
appdirs
pytest
